Al generar de plantillas deja números como números --y no como cadenas. Closes https://gitlab.com/pasosdeJesus/cor1440_gen/-/issues/49

+++ lib/heb412_gen/concerns/controllers/plantillahcm_controller.rb
@@ -216,7 +216,7 @@ module Heb412Gen
                 else
                   r[c.nombrecampo.to_sym]
                 end
-                unless v.is_a?(Integer)
+                unless v.is_a?(Numeric)
                   v = v.to_s
                 end
